+++ rest/private/orders/get-order-trigger-histories.go
@@ -35,8 +35,8 @@ type TriggerHistory struct {
 	TrailStart   float64 `json:"trailStart"`
 	TrailValue   float64 `json:"trailValue"`
 
-	ID               int       `json:"id"`
-	OrderID          int       `json:"orderId"`
+	ID               int64     `json:"id"`
+	OrderID          int64     `json:"orderId"`
 	Size             int       `json:"size"`
 	FilledSize       int       `json:"filledSize"`
 	RetryUntilFilled bool      `json:"retryUntilFilled"`
